requirements.txt
ebooklib
langdetect
transformers
requests
pypdf
python-docx
beautifulsoup4
psutil
reportlab
